Get Orbach set up to run under a WSGI standalone container.
Twisted and GUnicorn need to be able to just import the module and
access the app variable (which is the WSGI callable).  Meanwhile, the
development run script needs to be able to do some configuration on the
app before starting it.  By setting up (but not running) the app on
module import, we can pass it to a WSGI container and have our
development script modify the app before starting it.