Placate Nose load errors due to missing SQLAlchemy attributes.
It appears that Nose tries to load all modules it can find to search
for unit tests.  With the SQLAlchemy declarative style, however, the
orbach.model module would not load correctly because there was no
SQLAlchemy context.  Telling Nose not to load the application classes
with "tests=test" in setup.cfg works but then breaks Nose's ability to
zero-in on specific tests on the command-line.

This patch provides a better fix.  If no SQLAlchemy declarative base
exists, we create a stub one.  Nose can now load the modules.