Fix more issues with Flask demanding things at module scope.
Flask-RESTful does not allow Blueprints to be registered multiple times
(despite the Flask docs stating that doing so is acceptable).  This
requirement lead to the discovery that the orbach.app attribute was not
getting reset after every test.  The Flask docs annoyingly only provide
examples where everything is done at module scope and when "app" exists
at the module scope, it doesn't get destroyed at the end of a test.
This patch moves the app object out of module scope.